# openapi.yaml
openapi: 3.0.0
info:
  title: TiansuoVM API
  description: TiansuoVM平台API文档
  version: 1.0.0
servers:
  - url: /api/v1
    description: TiansuoVM API基础路径
...
paths:
...
  /vms/{id}:
    get:
      summary: 获取虚拟机详情
      tags:
        - 虚拟机管理
      operationId: getVM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
    delete:
      summary: 标记删除虚拟机
      tags:
        - 虚拟机管理
      operationId: deleteVM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
  /vms/{id}/recover:
    post:
      summary: 恢复已标记删除的虚拟机
      tags:
        - 虚拟机管理
      operationId: recoverVM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
  /vms/{id}/start:
    post:
      summary: 启动虚拟机
      tags:
        - 虚拟机管理
      operationId: startVM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
...
  /vms/{id}/restart:
    post:
      summary: 重启虚拟机
      tags:
        - 虚拟机管理
      operationId: restartVM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string

  # 镜像管理API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 